+++ src/day_1/example.txt
@@ -1,0 +1,10 @@
+199
+200
+208
+210
+200
+207
+240
+269
+260
+263
