+++ src/day_2/example.txt
@@ -1,0 +1,6 @@
+forward 5
+down 5
+forward 8
+up 3
+down 8
+forward 2
