+++ src/day_6/example.txt
@@ -1,0 +1,1 @@
+3,4,3,1,2
