+++ src/day_7/example.txt
@@ -1,0 +1,1 @@
+16,1,2,0,4,2,7,1,2,14
