--- src/day_8/part_1.rs
@@ -1,3 +1,0 @@
-pub fn find_solution() -> u32 {
-    0
-}
